chore: use types over strings for events
